Refactor TokenVisitor into smaller focused utility components
- Extract TriviaProcessor for trivia conversion logic
- Extract SyntaxClassifier for node classification and name cleaning
- Extract StructureExtractor for complex structure analysis
- Extract NodeFactory for TreeNode creation
- Reduce TokenVisitor from 357 to ~200 lines using composition
- Reduce visitPre() method from 134 to ~25 lines
- Reduce processTriviaPiece() method from 50 to 3 lines
- Maintain all existing functionality and public API
- All tests pass, no breaking changes
